third_party/blink/web_tests/external/wpt/webhid/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 112 tests; 63 PASS, 49 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS idl_test setup
PASS idl_test validation
PASS Partial interface Navigator: original interface defined
PASS Partial interface Navigator: member names are unique
PASS Partial interface mixin NavigatorID: member names are unique
PASS Navigator includes NavigatorID: member names are unique
PASS Navigator includes NavigatorLanguage: member names are unique
PASS Navigator includes NavigatorOnLine: member names are unique
PASS Navigator includes NavigatorContentUtils: member names are unique
PASS Navigator includes NavigatorCookies: member names are unique
PASS Navigator includes NavigatorPlugins: member names are unique
PASS Navigator includes NavigatorConcurrentHardware: member names are unique
PASS HID interface: existence and properties of interface object
PASS HID interface object length
PASS HID interface object name
PASS HID interface: existence and properties of interface prototype object
PASS HID interface: existence and properties of interface prototype object's "constructor" property
PASS HID interface: existence and properties of interface prototype object's @@unscopables property
PASS HID interface: attribute onconnect
PASS HID interface: attribute ondisconnect
PASS HID interface: operation getDevices()
PASS HID interface: operation requestDevice(HIDDeviceRequestOptions)
PASS HID must be primary interface of navigator.hid
PASS Stringification of navigator.hid
PASS HID interface: navigator.hid must inherit property "onconnect" with the proper type
PASS HID interface: navigator.hid must inherit property "ondisconnect" with the proper type
PASS HID interface: navigator.hid must inherit property "getDevices()" with the proper type
PASS HID interface: navigator.hid must inherit property "requestDevice(HIDDeviceRequestOptions)" with the proper type
PASS HID interface: calling requestDevice(HIDDeviceRequestOptions) on navigator.hid with too few arguments must throw TypeError
PASS HIDConnectionEvent interface: existence and properties of interface object
PASS HIDConnectionEvent interface object length
PASS HIDConnectionEvent interface object name
PASS HIDConnectionEvent interface: existence and properties of interface prototype object
PASS HIDConnectionEvent interface: existence and properties of interface prototype object's "constructor" property
PASS HIDConnectionEvent interface: existence and properties of interface prototype object's @@unscopables property
PASS HIDConnectionEvent interface: attribute device
PASS HIDInputReportEvent interface: existence and properties of interface object
FAIL HIDInputReportEvent interface object length assert_equals: wrong value for HIDInputReportEvent.length expected 2 but got 0
PASS HIDInputReportEvent interface object name
PASS HIDInputReportEvent interface: existence and properties of interface prototype object
PASS HIDInputReportEvent interface: existence and properties of interface prototype object's "constructor" property
PASS HIDInputReportEvent interface: existence and properties of interface prototype object's @@unscopables property
PASS HIDInputReportEvent interface: attribute device
PASS HIDInputReportEvent interface: attribute reportId
PASS HIDInputReportEvent interface: attribute data
FAIL HIDReportItem interface: existence and properties of interface object assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface object length assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface object name assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: existence and properties of interface prototype object assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute isAbsolute assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute isArray assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute isRange assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute hasNull assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute usages assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute usageMinimum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute usageMaximum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute designatorMinimum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute designatorMaximum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute stringMinimum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute stringMaximum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute reportSize assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute reportCount assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute unitExponent assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute unit assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute logicalMinimum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute logicalMaximum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute physicalMinimum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportItem interface: attribute physicalMaximum assert_own_property: self does not have own property "HIDReportItem" expected property "HIDReportItem" missing
FAIL HIDReportInfo interface: existence and properties of interface object assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDReportInfo interface object length assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDReportInfo interface object name assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDReportInfo interface: existence and properties of interface prototype object assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDReportInfo interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDReportInfo interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDReportInfo interface: attribute reportId assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDReportInfo interface: attribute items assert_own_property: self does not have own property "HIDReportInfo" expected property "HIDReportInfo" missing
FAIL HIDCollectionInfo interface: existence and properties of interface object assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface object length assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface object name assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: existence and properties of interface prototype object assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: attribute usagePage assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: attribute usage assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: attribute children assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: attribute inputReports assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: attribute outputReports assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: attribute featureReports assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: attribute reportIds assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: operation getField(BufferSource, HIDFieldOptions) assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
FAIL HIDCollectionInfo interface: operation setField(BufferSource, HIDFieldOptions, double) assert_own_property: self does not have own property "HIDCollectionInfo" expected property "HIDCollectionInfo" missing
PASS HIDDevice interface: existence and properties of interface object
PASS HIDDevice interface object length
PASS HIDDevice interface object name
PASS HIDDevice interface: existence and properties of interface prototype object
PASS HIDDevice interface: existence and properties of interface prototype object's "constructor" property
PASS HIDDevice interface: existence and properties of interface prototype object's @@unscopables property
PASS HIDDevice interface: attribute oninputreport
PASS HIDDevice interface: attribute opened
PASS HIDDevice interface: attribute vendorId
PASS HIDDevice interface: attribute productId
PASS HIDDevice interface: attribute productName
PASS HIDDevice interface: attribute collections
PASS HIDDevice interface: operation open()
PASS HIDDevice interface: operation close()
PASS HIDDevice interface: operation sendReport(octet, BufferSource)
PASS HIDDevice interface: operation sendFeatureReport(octet, BufferSource)
PASS HIDDevice interface: operation receiveFeatureReport(octet)
PASS Navigator interface: attribute hid
PASS Navigator interface: navigator must inherit property "hid" with the proper type
Harness: the test ran to completion.
